Added unit test for default x5c behavior. Clean up.
